requirements.txt
python-telegram-bot>=21.1
asyncio>=3.4.3
SQLAlchemy>=2.0.30
psycopg2>=2.9.9